Slurm backend: no max pending job limit any more.
This feature made the use of `minstructor` in combination with slurm
less efficient, as it was not trivial to submit jobs to different
slurm partitions, which can theoretically be executed in parallel.
But if there is a maximum number of pending jobs per user, two
simultaneously running minstructor instances, which schedule two
different partitions, basically compete with each other, and it might
happen that only one partition is scheduled at one moment.