Feature: backend argument expansion
The expansion of the command patterns is now
also applied to the backend argument string from the CLI.
Previously one had to duplicate a `minstructor` command
for different backend configurations, which resulted usually
in a copy paste solution of the minstructor command. An
examples has been added to the man page.